In progress commit, add all Conda Python 3 widget type templates
Conda Python 3 is a language that can now show up as in option in
dashboard, so there should be templates for all default charts widgets
for Conda Python 3. However, in this case, inputs and output will be
fixed for outputting and displaying an image.

This is a bit HACKY and ugly to implement due to the abstractions done
previously. But a better way is unclear without removing the scijava
discovery abstractions for the easy creation of new widget types.