BDV export:replace JDialog with TransformedBoxSelectionDialog
Use the TransformedBoxSelectionDialog for selection of the region for
export to ImagePlus in BDV. This allow direct selection in DBV and in a
BDV dialog. If rover sync is off the selection region is in the middle
of the current view. Otherwise, it is centered around the molecule based
on the Location settings.