Mars Bdv Commands: one active window at a time, bring to front if open
When the MarsBdvCommand dialogs are closed, sometimes they are not
garbage collected right away. Now we check if a JDialog exists and is
visible and if so we bring it to the front. Otherwise, if it is not
visible we create a new one. The next cycle of garbage collection should
eventually remove all old JDialogs.

This ensures only one window is open at a time for each command.