Add synchronized Bdv windows activation option
There has always been an option to open several Bdv frames at once that are synchronized with each other, so that channels can be turned on and off and different views can be used. But no one uses it, so now the option isn't offered by default, but can be activated in the settings.