BdvViewTable: Do not show image path selection dialog on creation
BdvSource are created using the plus button that appear when typing at the bottom of Bdv list. Clicking the plus button no longer brings up the path selection dialog. Instead paths need to be added in the BdvSourceOptionsPane using the Browse button. This make it faster and easier to add sources and copy paths from other records. This also removed redundant code.