WIP chartfx to 11.3 migrate to style property of dataset
With the update to 11.3.1 many style properties of datasets are now stored in the style string and created as correct JavaFX CSS using DataSetStyleBuilder. This update migrates as much as possible to this format. Scatter points are weird and required using MarkerLineWidth for MarkerSize and directly setting fill from MarkerColor in Renderer. Could be bugs in chartfx style conversion.