WIP add swing menubar
With the jump from JavaFX 8 to 21 focus issues with the Stage in a swing application appeared. But putting everything in a JFXPanel in a JFrame also bring the same problems as in early versions. We moved away from that approach several years earlier.

This commit extends a JFrame wrapper for the Stage to include a swing menubar. This prevents triggers of Fiji menu shortcuts even when the stage has focus, which appeared as a new issue with the migration to JavaFX 21 and above.